experiment/requirements.txt
paddleocr[doc-parser]
safetensors
gradio>=5.0,<6.0
pillow
